Delete danceandlife.html
Since I am creating a new branch(assignment 2)  the old repository, and the new branch shows the same content/files as the old branch. I decided to delete all the files from the new branch and add the assignment 2 file from my local machine and commit. That way, Git hub will be able to show assignment 2 in a new branch.